Create type-preposition.js
