+++ src/view/point-editor/type-preposition.js
@@ -1,0 +1,17 @@
+import {
+  toFirstUpperCase,
+} from '../../utils/utils';
+
+import {
+  pointTypeToPreposition,
+} from './data';
+
+const getPointTypeWithPreposition = (currentType) => {
+  const printType = toFirstUpperCase(currentType);
+  const preposition = pointTypeToPreposition[currentType];
+  return preposition ? `${printType} ${preposition}` : ``;
+};
+
+export {
+  getPointTypeWithPreposition,
+};
